fix(decks-builder): check applies the validator's spec rules with its own patterns
PROMPTS-CONTRACTS-03 / VALIDATOR-CODE-5: check passed specs the build then
failed on rules it could read in the spec. nbg_spec now imports EM_DASH,
EN_DASH, SOURCE_AS_OF, ALT_TEXT_PLACEHOLDER, ALT_TEXT_LEAD_IN and fold from
nbg_validate, so a change to either side (a widened as-of test, say)
reaches both.

- An em dash or a typed ' -- ' in drawn text is an error (was a warning);
  in speaker notes, alt text or storyline fields, which nothing draws, it
  stays a warning. A spaced en dash is a warning, as in the validator.
- An exhibit (a chart or table anywhere on the slide) whose source line has
  no year or date is an error; elsewhere a warning. The whole line counts:
  a year in the name, as_of or basis dates it.
- Alt text that is a filename or an autoname, opens with 'image of', or
  repeats text on the slide is an error.
- x- keys are skipped by the text walk: pipeline notes are never checked.

Also types the doughnut legend band height as float, for mypy.

+++ plugins/decks/tools/nbg-presentation/nbg_chart.py
@@ -757,7 +757,7 @@ def _legend_band(names: list[str], width: float) -> float:
         if used and used + entry > width - 0.2:
             rows, used = rows + 1, 0.0
         used += entry
-    return rows * m.line_height(size) + 0.15
+    return float(rows * m.line_height(size) + 0.15)
 
 
 def _ring_square(
